upgrade wgpu. work around new wgpu lifetimes (this was painful)

+++ src/render/render_graph/uniform.rs
@@ -171,7 +171,7 @@ pub struct UniformInfo<'a> {
 }
 
 pub struct DynamicUniformBufferInfo {
-    pub offsets: HashMap<Entity, u64>,
+    pub offsets: HashMap<Entity, u32>,
     pub capacity: u64,
     pub count: u64,
 }
